move rFullCircle into features/map
the map is now the only page that uses the rFullCircle selector and associated code. so this PR removes references to it from all other parts of the app and adds a fetchCircle query when the map page loads

remove the useFixCircleState hook; it's no longer needed

move the remaining recoil state into recoilState/index
