+++ src/features/auth/useLogout.ts
@@ -1,7 +1,7 @@
 import { client } from 'lib/gql/client';
 
 import { useRecoilLoadCatch } from 'hooks';
-import { rApiManifest, rApiFullCircle } from 'recoilState/db';
+import { rApiManifest } from 'recoilState';
 
 import { useSavedAuth } from './useSavedAuth';
 
@@ -11,7 +11,6 @@ export const useLogout = (remote = false) => {
     ({ set }) =>
       async () => {
         clearSavedAuth();
-        set(rApiFullCircle, new Map());
         set(rApiManifest, undefined);
 
         if (remote)
