+++ src/hooks/migration.ts
@@ -3,55 +3,9 @@
 //
 // If at all possible, write new code that uses Recoil only in this file.
 
-import { useEffect, useState } from 'react';
+import { useRecoilValue } from 'recoil';
 
-import { useRecoilValue, useRecoilState, useRecoilValueLoadable } from 'recoil';
-
-import { DebugLogger } from '../common-lib/log';
-import { useFetchCircle } from 'hooks/legacyApi';
-import {
-  rSelectedCircleId,
-  rApiManifest,
-  rManifest,
-  rApiFullCircle,
-  rSelectedCircleIdSource,
-} from 'recoilState';
-
-const logger = new DebugLogger('hooks/migration');
-
-// if you have a new page that doesn't use Recoil and is related to a specific
-// circle, you may want to use this hook to make sure that if you then navigate
-// away to a legacy page, that new page shows the correct circle.
-export const useFixCircleState = (circleId: number | undefined) => {
-  const recoilCircleId = useRecoilValueLoadable(rSelectedCircleId).valueMaybe();
-  const fullCircles = useRecoilValue(rApiFullCircle);
-  const [, setCircleIdSource] = useRecoilState(rSelectedCircleIdSource);
-  const fetchCircle = useFetchCircle();
-  logger.log(`useFixCircleState, circle id: ${circleId}`);
-
-  const [ready, setReady] = useState(false);
-
-  useEffect(() => {
-    if (!circleId) {
-      setReady(true);
-      return;
-    }
-
-    if (circleId === recoilCircleId) {
-      logger.log(`circle ids match`);
-      setReady(true);
-    } else if (fullCircles.has(circleId)) {
-      logger.log(`reusing circle data: ${circleId}`);
-      setCircleIdSource(circleId);
-      setReady(true);
-    } else {
-      logger.log(`fetching circle data: ${circleId}`);
-      fetchCircle({ circleId, select: true }).then(() => setReady(true));
-    }
-  }, [circleId, recoilCircleId]);
-
-  return ready;
-};
+import { rManifest } from 'recoilState';
 
 export const useRoleInCircle = (circleId: number | undefined) => {
   const manifest = useRecoilValue(rManifest);
@@ -67,12 +21,6 @@ export const useCanVouch = (circleId: number) => {
   return !(user?.non_giver && circle?.only_giver_vouch) && circle?.hasVouching;
 };
 
-export const useSomeCircleId = () => {
-  const selectedId = useRecoilValueLoadable(rSelectedCircleId).valueMaybe();
-  const firstId = useRecoilValue(rApiManifest)?.myUsers[0]?.circle_id;
-  return selectedId ?? firstId;
-};
-
 export const useShowGive = (circleId: number) => {
   const manifest = useRecoilValue(rManifest);
   const circle = manifest.circles.find(c => c.id === circleId);
@@ -83,7 +31,5 @@ export const useShowGive = (circleId: number) => {
 };
 
 export const useIsInCircle = (circleId: number) => {
-  const manifest = useRecoilValue(rManifest);
-  const user = manifest.myProfile.myUsers.find(u => u.circle_id === circleId);
-  return !!user?.role;
+  return useRoleInCircle(circleId) !== undefined;
 };
