+++ src/hooks/useApiAdminCircle.ts
@@ -9,7 +9,6 @@ import { useQueryClient } from 'react-query';
 
 import { fileToBase64 } from '../lib/base64';
 import { ValueTypes } from '../lib/gql/__generated__/zeus';
-import { useFetchCircle } from 'hooks/legacyApi';
 import { QUERY_KEY_GET_MEMBERS_PAGE_DATA } from 'pages/MembersPage/getMembersPageData';
 
 import { useRecoilLoadCatch } from './useRecoilLoadCatch';
@@ -60,14 +59,12 @@ export const adminUpdateUser = async (
 };
 
 export const useApiAdminCircle = (circleId: number) => {
-  const fetchCircle = useFetchCircle();
   const queryClient = useQueryClient();
 
   const updateCircleLogo = useRecoilLoadCatch(
     () => async (newLogo: File) => {
       const image_data_base64 = await fileToBase64(newLogo);
       await mutations.updateCircleLogo(circleId, image_data_base64);
-      await fetchCircle({ circleId });
     },
     [circleId]
   );
@@ -76,7 +73,6 @@ export const useApiAdminCircle = (circleId: number) => {
     () => async (params: ValueTypes['CreateEpochInput']['params']) => {
       params.time_zone = Intl.DateTimeFormat().resolvedOptions().timeZone;
       await mutations.createEpoch({ circle_id: circleId, params });
-      await fetchCircle({ circleId });
     },
     [circleId],
     { hideLoading: false }
@@ -103,7 +99,6 @@ export const useApiAdminCircle = (circleId: number) => {
           circle_id: circleId,
           description,
         });
-        await fetchCircle({ circleId });
       },
     [circleId],
     { hideLoading: false }
@@ -119,7 +114,6 @@ export const useApiAdminCircle = (circleId: number) => {
         }
       ) => {
         await mutations.updateActiveRepeatingEpoch(circleId, epochId, params);
-        await fetchCircle({ circleId });
       },
     [circleId],
     { hideLoading: false }
@@ -128,7 +122,6 @@ export const useApiAdminCircle = (circleId: number) => {
   const deleteEpoch = useRecoilLoadCatch(
     () => async (epochId: number) => {
       await mutations.deleteEpoch(circleId, epochId);
-      await fetchCircle({ circleId });
     },
     [circleId]
   );
@@ -136,7 +129,6 @@ export const useApiAdminCircle = (circleId: number) => {
   const updateUser = useRecoilLoadCatch(
     () => async (userAddress: string, params: UpdateUsersParam) => {
       await adminUpdateUser(circleId, userAddress, params);
-      await fetchCircle({ circleId });
     },
     [circleId]
   );
@@ -145,9 +137,6 @@ export const useApiAdminCircle = (circleId: number) => {
     () => async (userAddress: string) => {
       await mutations.deleteUser(circleId, userAddress);
       await queryClient.invalidateQueries(QUERY_KEY_GET_MEMBERS_PAGE_DATA);
-
-      // probably unnecessary now
-      await fetchCircle({ circleId });
     },
     [circleId]
   );
@@ -174,7 +163,6 @@ export const useApiAdminCircle = (circleId: number) => {
   const restoreCoordinape = useRecoilLoadCatch(
     () => async (circleId: number) => {
       await mutations.restoreCoordinapeUser(circleId);
-      await fetchCircle({ circleId });
     }
   );
