+++ src/pages/MapPage/index.tsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react';
 
 import { AMDrawer } from 'features/map/AMDrawer';
 import { AMForceGraph } from 'features/map/AMForceGraph';
+import { useFetchCircle } from 'features/map/queries';
 import { useSetAmEgoAddress } from 'features/map/state';
 import { ThemeContext } from 'features/theming/ThemeProvider';
 import { useNavigate, useLocation } from 'react-router-dom';
@@ -9,6 +10,7 @@ import { useRecoilCallback } from 'recoil';
 
 import { Box } from '../../ui';
 import { rDevMode } from 'recoilState';
+import { useCircleIdParam } from 'routes/hooks';
 
 const MAP_HIGHLIGHT_PARAM = 'highlight';
 
@@ -16,6 +18,14 @@ export default function MapPage() {
   const location = useLocation();
   const navigate = useNavigate();
   const setAmEgoAddress = useSetAmEgoAddress();
+  const fetchCircle = useFetchCircle();
+  const circleId = useCircleIdParam();
+
+  useEffect(() => {
+    (async () => {
+      await fetchCircle({ circleId, select: true });
+    })();
+  }, []);
 
   useEffect(() => {
     const queryParams = new URLSearchParams(location.search);
